+++ src/components/Pagination/examples/index.js
@@ -1,0 +1,14 @@
+/* eslint-disable react/jsx-key */
+import React from 'react';
+import Pagination from '../index';
+
+export default [
+  {
+    examples: [
+      <Pagination
+        currentPage={ 1 }
+        totalPages={ 100 }
+      />,
+    ],
+  },
+];