+++ src/main.js
@@ -17,6 +17,7 @@ export { default as Input } from './components/Input';
 export { default as Label } from './components/Label';
 export { default as Link } from './components/Link';
 export { default as OverlayWrapper } from './components/OverlayWrapper';
+export { default as Pagination } from './components/Pagination';
 export { default as ProgressBar } from './components/ProgressBar';
 export { default as ProgressDots } from './components/ProgressDots';
 export { default as Popover } from './components/Popover';