+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 ## Unreleased
 - [Patch] Update Travis.yml to send alerts to #oui-alerts in Slack
 - [Patch] Add hover and focus error states to input
+- [Feature] Create Pagination component
 
 ## 40.5.0 - 2018-10-10
 - [Patch] Font-weight 400 for tiny button